requirements.txt
asgiref==3.8.1
asyncio==3.4.3
django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6